pyproject.toml
[build-system]
requires = ["setuptools","wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mal4py"
version = "0.1.5"
authors = [
    { name="Tony White", email="kirock1195+dev@gmail.com" },
]
maintainers = [
    { name="antoblanco", email="kirock1195+dev@gmail.com"},
]
keywords = ["api", "myanimelist", "MAL", "Anime", "Manga"]
description = "MAL API An unofficial MyAnimeList API for Python 3"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.9"
dependencies = [
    "requests>=2.25.0",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/AntoBlanco/MAL4Py"
"Bug Tracker" = "https://github.com/AntoBlanco/MAL4Py/issues"
[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "python-dotenv",
    "black"
]